requirements.txt
playwright
pytest
pytest-asyncio
python-dotenv